fix: scope reward ad helpers to android
