+++ core/ui/src/androidHostTest/kotlin/me/matsumo/fanbox/core/ui/ads/RewardAdShowStateTest.kt
@@ -1,6 +1,6 @@
 package me.matsumo.fanbox.core.ui.ads
 
-import kotlin.test.Test
+import org.junit.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
