+++ core/ui/src/androidMain/kotlin/me/matsumo/fanbox/core/ui/ads/RewardAdShowResult.kt
@@ -1,0 +1,10 @@
+package me.matsumo.fanbox.core.ui.ads
+
+import androidx.compose.runtime.Immutable
+
+/** リワード広告の表示完了結果。 */
+@Immutable
+data class RewardAdShowResult(
+    val requestId: Long,
+    val isRewardEarned: Boolean,
+)
